Example household survey data

Example of a household survey dataset, used in unit tests and vignettes for the networkreporting package.

Usage

example.survey

Format

A data frame with 2,406 rows and 36 variables:

id

a unique identifier

cluster

the cluster (part of the complex survey design)

region

the region (part of the complex survey design)

indweight

the individual weight (relative)

sex

the sex of the respondent

age.cat

the age category of the respondent

widower

reported connections to widowers

nurse.or.doctor

reported connections to nurses or doctors

male.community.health

reported connections to male community health workers

teacher

reported connections to teachers

woman.smoke

reported connections to women who smoke

priest

reported connections to priests

civil.servant

reported connections to civil servants

woman.gave.birth

reported connections to women who gave birth

muslim

reported connections to Muslims

incarcerated

reported connections to people who are incarcerated

judge

reported connections to people who are judges

man.divorced

reported connections to men who are divorced

treatedfortb

reported connections to men who are divorced

nsengimana

reported connections to Nsegimanas

murekatete

reported connections to Murekatetes

twahirwa

reported connections to Twahirwas

mukandekezi

reported connections to Mukandekezis

nsabimana

reported connections to Nsabimanas

mukamana

reported connections to Mukamanas

ndayambaje

reported connections to Ndayambajes

nyiraneza

reported connections to Nyiranezas

bizimana

reported connections to Bizimanas

nyirahabimana

reported connections to Nyirahabimanas

ndagijimana

reported connections to Ndagijimanas

mukandayisenga

reported connections to Mukandayisengas

died

reported connections to people who died

sex.workers

reported connections to sex workers

msm

reported connections to msm

idu

reported connections to injecting drug users

clients

reported connections to people who are clients of sex workers