make an age.groups object

TODO – would be nice if this could take either widths or bins / breaks

Usage

make.age.groups(start, widths, names)

Arguments

start

the first age

widths

the widths of the subsequent age groups

names

the names of the age groups

Value

an object (list) with the widths, names, and number of age groups, as well as a matrix called template which has the start and end of each age interval. the intervals in template are close on the left but not the right; eg, exp.start of 10 and exp.end of 20 means [10, 20) in terms of exact ages (this is useful later on, for making individual age schedules based on, eg, birth dates)